fix alottta shit with pablo
