+++ routes/account.js
@@ -4,69 +4,71 @@ const bcrypt = require('bcrypt')
 const session = require('express-session')
 const LocalStrategy = require('passport-local').Strategy
 const router = express.Router()
-const knexConfig = require('../knexfile')
-const knex = require('knex')(knexConfig.development)
 const app = express()
 app.use(express.json())
 app.use(express.urlencoded({ extended: false }))
 
-// passport configuration
-passport.use(new LocalStrategy(
-  function (username, password, done) {
-    console.log('checking to see if there is a user...')
-    // is there an exisitng use with that username? see if there a username that matches the username
-    findUserByUsernameStrategy(username)
-      .then(function (result) {
-        console.log('Heres what the query found in the db: ', result)
-        const user = result[0]
-        console.log('Now we compare user.password: ' + user.password + ' with password: ' + password)
-        bcrypt.compare(password, user.password)
-          .then(function (bool) {
-            if (user && bool) {
-              console.log('Password matched! ✨')
-              return done(null, user)
-            } else {
-              console.log('Password did not match! 🤮')
-              return done(null, false)
-            }
-          })
-      })
-      .catch(function (error) {
-        console.log('findUserByUsernameStrategy error: ', error)
-        return done(error)
-      })
-  }
-))
+// // passport configuration
+// passport.use(new LocalStrategy(
+//   function (username, password, done) {
+//     console.log('🔸 Checking to see if there is a user with that username...')
+//     findUserByUsernameStrategy(username)
+//       .then(function (result) {
+//         if (result) {
+//           console.log('✅ User found! This is the user: ', result)
+//           const user = result
+//           console.log('🔸 Comparing the passwords: ' + user.password + ' 🆚 ' + password)
+//           bcrypt.compare(password, user.password)
+//             .then(function (bool) {
+//               if (user && bool) {
+//                 console.log('Password matched! ✨')
+//                 return done(null, user)
+//               } else {
+//                 console.log('Password did not match! 🤮')
+//                 return done(null, false)
+//               }
+//             })
+//         } else {
+//           console.log('Theres no user with that login')
+//           return done(null, false)
+//         }
+//       })
+//       .catch(function (error) {
+//         console.log('findUserByUsernameStrategy error: ', error)
+//         return done(error)
+//       })
+//   }
+// ))
 
-passport.serializeUser(function (user, done) {
-  console.log('serializing user....')
-  done(null, user.userId)
-})
+// passport.serializeUser(function (user, done) {
+//   console.log('👉🏻 SERIALIZING')
+//   done(null, user.userId)
+// })
 
-passport.deserializeUser(function (id, done) {
-  console.log('deserializing user....')
-  findUserByIdStrategy(id)
-    .then(function (user) {
-      done(null, user[0])
-    })
-    .catch(function (error) {
-      console.log('deserializeUser err:', error)
-      done(error, null)
-    })
-})
+// passport.deserializeUser(function (id, done) {
+//   console.log('👉🏻 DESERIALIZING')
+//   findUserByIdStrategy(id)
+//     .then(function (user) {
+//       done(null, user[0])
+//     })
+//     .catch(function (error) {
+//       console.log('deserializeUser err:', error)
+//       done(error, null)
+//     })
+// })
 
-// session configuration
-app.set('trust proxy', 1) // trust first proxy
-app.use(session({
-  secret: 'keyboard cat',
-  resave: false,
-  saveUninitialized: true,
-  cookie: {}
-}))
+// // session configuration
+// app.set('trust proxy', 1)
+// app.use(session({
+//   secret: process.env.SESSION_SECRET,
+//   resave: false,
+//   saveUninitialized: true,
+//   cookie: { secure: false }
+// }))
 
-// init passport
-app.use(passport.initialize())
-app.use(passport.session())
+// // init passport
+// app.use(passport.initialize())
+// app.use(passport.session())
 
 // routes below are prepended with /account from mounting on app.js
 router.get('/login', function (req, res) {
@@ -74,12 +76,12 @@ router.get('/login', function (req, res) {
 })
 
 router.post('/login', passport.authenticate('local', { failureRedirect: 'login', successRedirect: '/' }), function (req, res, next) {
-  console.log('posting to login')
+  console.log('✅Verifying login information and posting to /login...')
   if (validLoginInformation(req.body)) {
-    console.log('verifying input data...')
+    console.log('✅Login information is valid! Searching for the user in the database...')
     findUserByUsername(req.body)
       .then(function (userObj) {
-        console.log('This is after pulling from the db: ', userObj)
+        console.log('✅Found this user in the database: ', userObj)
         if (!userObj) {
           res.send('There was a login error!')
         } else {
@@ -143,6 +145,11 @@ router.post('/edit', function (req, res) {
   res.send('Change account settings')
 })
 
+router.get('/logout', function(req, res){
+  req.logout();
+  res.redirect('/');
+});
+
 // validation functions
 // --------------------------------------------------------------------------------------------------------
 function validRegisterInformation (user) {
@@ -159,43 +166,51 @@ function validLoginInformation (user) {
   return validUsername && validPassword
 }
 
-// knex queries
-// --------------------------------------------------------------------------------------------------------
-function findUserByUsernameStrategy (username) {
-  return knex('User').where('username', username)
-}
+// // knex queries
+// // --------------------------------------------------------------------------------------------------------
+// function findUserByUsernameStrategy (username) {
+//   return knex('User').where('username', username)
+//     .then(function (result){
+//       const user = result[0]
+//       return user
+//     })
+//     .catch(function(err){
+//       console.log(err)
+//       return null
+//     })
+// }
 
-function findUserByIdStrategy (id) {
-  return knex('User').where('userId', id)
-}
+// function findUserByIdStrategy (id) {
+//   return knex('User').where('userId', id)
+// }
 
-function findUserByUsername (user) {
-  const username = user.username
-  return knex.select().from('User').where({
-    username: username
-  })
-    .then(function (userArr) {
-      // const singleUser = userArr[0]
-      // console.log(userArr)
-      const singleUser = {
-        username: userArr[0].username,
-        password: userArr[0].password
-      }
-      console.log(singleUser)
-      return singleUser
-    })
-    .catch(function (err) {
-      console.error(err)
-    })
-}
+// function findUserByUsername (user) {
+//   const username = user.username
+//   return knex.select().from('User').where({
+//     username: username
+//   })
+//     .then(function (userArr) {
+//       // const singleUser = userArr[0]
+//       // console.log(userArr)
+//       const singleUser = {
+//         username: userArr[0].username,
+//         password: userArr[0].password
+//       }
+//       console.log(singleUser)
+//       return singleUser
+//     })
+//     .catch(function (err) {
+//       console.error(err)
+//     })
+// }
 
-function insertNewUser (user) {
-  return knex('User').insert({
-    fullName: user.fullName,
-    username: user.username,
-    email: user.email,
-    password: user.password
-  }, 'username')
-}
+// function insertNewUser (user) {
+//   return knex('User').insert({
+//     fullName: user.fullName,
+//     username: user.username,
+//     email: user.email,
+//     password: user.password
+//   }, 'username')
+// }
 
 module.exports = router // exports this router usually to app.js
